kernel: A bunch of small bug fixes

+++ kernel/sys/timer.h
@@ -3,9 +3,10 @@
 
 #if defined (__x86_64__)
 
+#include <stdint.h>
 #include <dev/lapic.h>
 
-static inline void sys_timer_oneshot(size_t us, void *function) {
+static inline void sys_timer_oneshot(uint64_t us, void *function) {
     lapic_timer_oneshot(us, function);
 }
